refactor: improve login prompt styling, footer padding, scroll-to-item and show plans for logged-out users
